requirements.txt
tqdm==4.48.2
tensorflow==2.6.2
numpy==1.17.4
gym==0.17.2
PyYAML==6.0.1
